+++ lib/src/languages/java/java11/java11_generator.dart
@@ -172,7 +172,7 @@ class ApolloCodeGeneratorJava11 extends ApolloCodeGenerator {
 
     for (var v in value.values) {
       if (v is ASTValueStringVariable) {
-        var s2 = generateASTValueStringVariable(v, '');
+        var s2 = generateASTValueStringVariable(v, '', null, list.isNotEmpty);
         list.add(s2);
       } else if (v is ASTValueStringExpresion) {
         var s2 = generateASTValueStringExpresion(v, '');
@@ -212,9 +212,15 @@ class ApolloCodeGeneratorJava11 extends ApolloCodeGenerator {
 
   @override
   StringBuffer generateASTValueStringVariable(ASTValueStringVariable value,
-      [String indent = '', StringBuffer? s]) {
+      [String indent = '', StringBuffer? s, bool precededByString = false]) {
     s ??= StringBuffer();
-    s.write('String.valueOf( ${value.variable.name} )');
+
+    if (precededByString) {
+      s.write(value.variable.name);
+    } else {
+      s.write('String.valueOf( ${value.variable.name} )');
+    }
+
     return s;
   }